Incorporate XeTeX functionality, at least in theory.
The XeTeX command still will not work.  This commit (and in fact this entire repo) is to ask for help on how to make this a reality.

The problem is that XeTeX does not produce the DVI files `dvips` obviously expects, which breaks the chain.  I don't know enough Perl to make this robust, or really even understand what's going on.